+++ package.json
@@ -25,7 +25,7 @@
     "cors": "^2.8.1",
     "express": "^4.14.1",
     "express-xml-bodyparser": "^0.3.0",
-    "fhir": "^3.2.0",
+    "fhir": "^3.3.0",
     "jsonwebtoken": "^8.0.1",
     "lodash": "^4.17.4",
     "moment": "^2.17.1",
